pyproject.toml
[project]
name = "xhs-assistant"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "httpx>=0.28.1",
    "mcp[cli]>=1.21.1",
    "langchain>=1.0.7",
    "langchain-openai>=1.0.3",
    "langchain-deepseek>=1.0.1",
    "langchain-core>=1.0.5",
    "python-dotenv>=1.0.0",
    "langgraph>=1.0.3",
]
